Move WindowsPathToCygwinPath to utils package instead of nodeps
