+++ pkg/util/utils.go
@@ -8,6 +8,7 @@ import (
 	"os"
 	osexec "os/exec"
 	"os/user"
+	"path/filepath"
 	"runtime"
 	"strings"
 	"time"
@@ -358,4 +359,16 @@ func ArrayToReadableOutput(slice []string) (response string, err error) {
 		return "", fmt.Errorf("empty slice")
 	}
 	return "[\n\t" + strings.Join(slice, "\n\t") + "\n]", nil
+}
+
+// WindowsPathToCygwinPath changes C:/path/to/something to //c/path/to/something
+// This is required for Docker bind mounts on Docker toolbox.
+// Sadly, if we have a Windows drive name, it has to be converted from C:/ to //c for Win10Home/Docker toolbox
+func WindowsPathToCygwinPath(windowsPath string) string {
+	windowsPath = filepath.ToSlash(windowsPath)
+	if string(windowsPath[1]) == ":" {
+		drive := strings.ToLower(string(windowsPath[0]))
+		windowsPath = "/" + drive + windowsPath[2:]
+	}
+	return windowsPath
 }
